modification sur les controller et impl de la classe agent

+++ src/main/java/com/fr/adaming/repository/AgentRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.fr.adaming.entity.Agent;
 
 @Repository
-public interface AgentRepository extends JpaRepository<Agent, Integer> {
+public interface AgentRepository extends JpaRepository<Agent, Long> {
 
 }
